dev-requirements.txt
# This file contains all the development requirements for our linting toolchain.
# Don't forget to pin your dependencies!
# This list will have to be migrated if you wish to use another dependency manager.

# Base tools
flake8~=4.0.1
isort~=5.10.1
pre-commit~=2.17.0
django~=4.2.2
djangorestframework~=3.14.0

# Flake8 plugins, see https://github.com/python-discord/code-jam-template/tree/main#plugin-list
flake8-docstrings~=1.6.0
